+++ .vscode/settings.json
@@ -1,19 +1,19 @@
 {
-  "workbench.colorCustomizations": {
-    "activityBar.activeBackground": "#08498b",
-    "activityBar.activeBorder": "#1e0210",
-    "activityBar.background": "#08498b",
-    "activityBar.foreground": "#e7e7e7",
-    "activityBar.inactiveForeground": "#e7e7e799",
-    "activityBarBadge.background": "#1e0210",
-    "activityBarBadge.foreground": "#e7e7e7",
-    "statusBar.background": "#05305b",
-    "statusBar.foreground": "#e7e7e7",
-    "statusBarItem.hoverBackground": "#08498b",
-    "titleBar.activeBackground": "#05305b",
-    "titleBar.activeForeground": "#e7e7e7",
-    "titleBar.inactiveBackground": "#05305b99",
-    "titleBar.inactiveForeground": "#e7e7e799"
-  },
-  "peacock.color": "#05305b"
+    "workbench.colorCustomizations": {
+        "activityBar.activeBackground": "#fadf5c",
+        "activityBar.activeBorder": "#059b81",
+        "activityBar.background": "#fadf5c",
+        "activityBar.foreground": "#15202b",
+        "activityBar.inactiveForeground": "#15202b99",
+        "activityBarBadge.background": "#059b81",
+        "activityBarBadge.foreground": "#e7e7e7",
+        "statusBar.background": "#f9d52a",
+        "statusBar.foreground": "#15202b",
+        "statusBarItem.hoverBackground": "#e9c207",
+        "titleBar.activeBackground": "#f9d52a",
+        "titleBar.activeForeground": "#15202b",
+        "titleBar.inactiveBackground": "#f9d52a99",
+        "titleBar.inactiveForeground": "#15202b99"
+    },
+    "peacock.color": "#f9d52a"
 }
