+++ src/utils/API.js
@@ -1,8 +1,8 @@
-const axios = require ("axios")
+const axios = require("axios")
 //local url communication
-//const URL_PREFIX = "http://localhost:3030"
+const URL_PREFIX = "http://localhost:3030"
 //When ready, the deployed site will use the following:
-const URL_PREFIX = "https://quiet-caverns-20153.herokuapp.com"
+// const URL_PREFIX = "https://quiet-caverns-20153.herokuapp.com"
 
 
 const API = {
@@ -15,7 +15,7 @@ const API = {
         console.log(userData)
         return axios.post(`${URL_PREFIX}/login`, userData)
     },
-    getAuthToken: token =>{
+    getAuthToken: token => {
         return axios.get(`${URL_PREFIX}/auth`, {
             headers: {
                 authorization: `Bearer: ${token}`
